// src/app/interfaces/sysrecord.ts
export interface Hospital {
    hospitalCode: string;
    hospitalName: string;
    hospitalShortName: string;
}
// ...
export interface PromotionGroupForHospital {
    id: number;
    promotionId: number;
    promotionNo: string;
    hospitalCode: string;
}

export interface Promotion {
    id: number;
    promotionNo: string;
    promotionName: string;
    promotionGroup: string;
    promotionGroupForHospital: PromotionGroupForHospital[];
    coverImageLocation: string;
    coverImageUrl: string;
    detailImageLocation: string;
    detailImageUrl: string;
    paStartDate: Date;
    paEndDate: Date;
    promotionTypeBrochure: string;
    promotionTypeRedeem: string;
    redeemQty: number;
    showByLocation: string;
    gender: string;
    nationalityCode: string;
    monthOfBirth: number;
    age: string;
    ageFrom: number;
    ageTo: number;
    provinceCode: string;
    memberType: string;
    memberNotRenewStatus: string;
    createBy: string;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
    updateLocation: string;
}

export interface PromotionGroupForHospital {
    id: number;
    promotionId: number;
    promotionNo: string;
    hospitalCode: string;
}


export interface MemberCardType {
    cardCode: string;
    cardDesc: string;
}

export interface Redeem {
    id: number;
    promoOrGiftId: number;
    redeemType: string;
    chivaCode: string;
    redeemDate: Date;
    redeemNo: string;
    promotionDetail: string;
    hospitalName: string;
    hospitalClaim: string;
    redeemCreateby: string;
    memoName: string;
    isClaim: boolean;
}